+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Deprecated]
 ### [Removed]
 ### [Fixed]
+- @aofl/cli-lib/git - fixed issue with archive command throwing an error when pulling multiple files.
 ### [Security]
 
 ---